[#2540] Move telemetry client classes into separate module
The current adapter client module is quite opinionated about its usage
scenario, i.e. it is intended to address requirements of protocol
adapters. However, other services also need to publish events but do not
need to query the registry not do they need to consume commands.

In order to better reflect such cases, the classes for sending Telemetry
and Event messages have been moved into dedicated telemetry,
telemetry-amqp and telemetry-kafka modules.
