+++ adapter-base/src/main/java/org/eclipse/hono/adapter/monitoring/ConnectionEventProducer.java
@@ -14,8 +14,8 @@ package org.eclipse.hono.adapter.monitoring;
 
 
 import org.eclipse.hono.adapter.client.registry.TenantClient;
-import org.eclipse.hono.adapter.client.telemetry.EventSender;
 import org.eclipse.hono.auth.Device;
+import org.eclipse.hono.client.telemetry.EventSender;
 
 import io.vertx.core.Future;
 import io.vertx.core.json.JsonObject;
