+++ adapter-base/src/test/java/org/eclipse/hono/adapter/monitoring/HonoEventConnectionEventProducerTest.java
@@ -22,10 +22,10 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import org.eclipse.hono.adapter.client.registry.TenantClient;
-import org.eclipse.hono.adapter.client.telemetry.EventSender;
 import org.eclipse.hono.adapter.monitoring.ConnectionEventProducer;
 import org.eclipse.hono.adapter.monitoring.HonoEventConnectionEventProducer;
 import org.eclipse.hono.auth.Device;
+import org.eclipse.hono.client.telemetry.EventSender;
 import org.eclipse.hono.util.EventConstants;
 import org.eclipse.hono.util.RegistrationAssertion;
 import org.eclipse.hono.util.ResourceLimits;
